+++ lib/src/reworked_bloc_widgets/bloc_listener.dart
@@ -7,8 +7,7 @@ import 'package:flutter_bloc/flutter_bloc.dart' as fb;
 /// Reworked version of [fb.BlocListener] which gets bloc not from context but
 /// by provided parameter
 /// {@endtemplate}
-class BlocListener<Bloc extends fb.StateStreamable<State>, State>
-    extends StatelessWidget {
+class BlocListener<B extends fb.StateStreamable<S>, S> extends StatelessWidget {
   /// {@macro bloc_listener}
   const BlocListener({
     required this.listener,
@@ -18,41 +17,21 @@ class BlocListener<Bloc extends fb.StateStreamable<State>, State>
     Key? key,
   }) : super(key: key);
 
-  /// {@template bloc_listener.listener}
-  /// Same as the 'listener' property for BlocListener from flutter_bloc
-  /// package.
-  ///
-  /// See the documentation there if you need.
-  /// {@endtemplate}
-  final fb.BlocWidgetListener<State> listener;
+  /// The [fb.BlocWidgetListener] which will be called on every state change.
+  final fb.BlocWidgetListener<S> listener;
 
-  /// {@template bloc_listener.bloc}
-  /// {@macro flutter_bloc_side_effect_listener_base.bloc}
-  /// Same as the 'bloc' property for BlocListener from flutter_bloc package.
-  ///
-  /// See the documentation there if you need.
-  /// {@endtemplate}
-  final Bloc bloc;
+  /// The [bloc] whose `state` will be listened to.
+  final B bloc;
 
-  /// {@template bloc_listener.listenWhen}
-  /// Same as the 'listenWhen' property for BlocListener from flutter_bloc
-  /// package.
-  ///
-  /// See the documentation there if you need.
-  /// {@endtemplate}
-  final fb.BlocListenerCondition<State>? listenWhen;
+  /// Optional condition to determine whether to call [listener].
+  final fb.BlocListenerCondition<S>? listenWhen;
 
-  /// {@template bloc_listener.child}
-  /// Same as the 'child' property for BlocListener from flutter_bloc
-  /// package.
-  ///
-  /// See the documentation there if you need.
-  /// {@endtemplate}
+  /// The widget which will be rendered as a descendant.
   final Widget? child;
 
   @override
   Widget build(BuildContext context) {
-    return fb.BlocListener<Bloc, State>(
+    return fb.BlocListener<B, S>(
       listener: listener,
       listenWhen: listenWhen,
       bloc: bloc,