+++ lib/src/side_effect_provider.dart
@@ -3,8 +3,7 @@ library flutter_bloc_side_effect;
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 /// An object that provides access to a stream of side effects over time.
-abstract class SideEffectProvider<SideEffect, State>
-    extends StateStreamable<State> {
-  /// The current [sideEffects] of side effects.
-  Stream<SideEffect> get sideEffects;
+abstract class SideEffectProvider<S, SE> extends StateStreamable<S> {
+  /// The stream of side effects.
+  Stream<SE> get sideEffects;
 }